Digital Elections | Kafka | PySpark | Redis | Streamlit/requirements.txt
psycopg2-binary
faker
redis
